LayoutTests/platform/chromium/http/tests/security/cross-frame-access-document-direct-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-document-direct-test-victim.html from frame with URL http://127.0.0.1:8000/security/resources/cross-frame-iframe-for-document-direct-test.html. Domains, protocols and ports must match.

Test cross-origin direct document access.



--------
Frame: 'theFrame'
--------
PASS: Cross origin direct document access denied.
